# inoa-measurement/src/main/resources/openapi/inoa-measurement-api.yaml
openapi: 3.0.3
info:
  title: Inoa Measurement API definition
  version: ${project.version}
  description: Definitions for Inoa Measurement.
  license:
    name: Apache License 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: GrayC GmbH
    url: https://grayc.de
servers:
  - url: http://measurement-service:8080
security:
  - Keycloak: []
  - BearerAuth: []
paths:
  /thing-types:
    get:
      tags:
        - thingTypes
      parameters:
        - $ref: '#/components/parameters/PageablePage'
        - $ref: '#/components/parameters/PageableSize'
        - $ref: '#/components/parameters/Sort'
        - name: filter
          description: 'Search in name. Supports wildcard *, is case insensitive.'
          in: query
          schema:
            type: string
            maxLength: 10
            example: dvh
      summary: 'Find thing types.'
      description: 'Returns all thing types.'
      operationId: findThingTypes
      responses:
        200:
          description: 'Returns thing types.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThingTypePage'
        401:
          description: 'Unauthorized.'
    post:
      tags:
        - thingTypes
      summary: 'Create thing type.'
      description: TODO
      operationId: createThingType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThingTypeCreate'
      responses:
        201:
          description: 'Tenant created.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThingType'
        400:
          description: 'Malformed.'
        401:
          description: 'Unauthorized.'
  /thing-types/{thing_type_id}:
    parameters:
      - $ref: '#/components/parameters/ThingTypeId'
    get:
      tags:
        - thingTypes
      summary: 'Find Thing type.'
      description: 'return a single thing type'
      operationId: findThingType
      responses:
        200:
          description: 'Thing type found.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThingType'
        401:
          description: 'Unauthorized.'
        404:
          description: 'Thing Type not found.'
    post:
      tags:
        - thingTypes
      summary: 'Create thing type.'
      description: TODO
      operationId: createThingTypeWithId
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThingTypeCreate'
      responses:
        201:
          description: 'Tenant created.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThingType'
        400:
          description: 'Malformed.'
        401:
          description: 'Unauthorized.'
    put:
      tags:
        - thingTypes
      summary: 'Update Thing type.'
      description: TODO
      operationId: updateThingType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThingTypeUpdate'
      responses:
        200:
          description: 'Thing type updated.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThingType'
        400:
          description: 'Malformed.'
        401:
          description: 'Unauthorized.'
        404:
          description: 'Thing type not found.'

    delete:
      tags:
        - thingTypes
      summary: 'Delete thing type.'
      description: TODO
      operationId: deleteThingType
      responses:
        204:
          description: 'Thing type deleted.'
        401:
          description: 'Unauthorized.'
        404:
          description: 'Thing type not found.'
  /thing-types/by-reference/{thing_type_reference}:
    parameters:
      - name: thing_type_reference
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - thingTypes
      summary: 'Find Thing type.'
      description: 'return a single thing type'
      operationId: findThingTypeByReference
      responses:
        200:
          description: 'Thing type found.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThingType'
        401:
          description: 'Unauthorized.'
        404:
          description: 'Thing Type not found.'
  /thing-types/{thing_type_id}/details:
    parameters:
      - $ref: '#/components/parameters/ThingTypeId'
    get:
      tags:
        - thingTypes
      summary: 'Find Thing type with its details.'
      description: 'Find Thing type with its details.'
      operationId: findThingTypeWithDetails
      responses:
        200:
          description: 'Thing type with details found.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThingTypeDetail'
        401:
          description: 'Unauthorized.'
        404:
          description: 'Thing Type not found.'
...
components:
  schemas:
    Timestamp:
      description: 'Common timestamp for created/updated timestamps.'
      type: string
      format: date-time
      example: '2021-01-01T00:00:00'
    Id:
      description: 'Id as technical reference (never changes).'
      type: string
      format: uuid
    ThingTypeId:
      description: 'Id as technical reference (never changes).'
      type: string
...
    Name:
      description: 'Name.'
      type: string
...
    ThingType:
      description: 'User with fields.'
      type: object
      properties:
        id:
          $ref: '#/components/schemas/ThingTypeId'
        name:
          $ref: '#/components/schemas/Name'
        thing_type_reference:
          description: 'thing_type_reference'
          type: string
        json_schema:
          description: 'json_schema'
          type: object
          additionalProperties: { }
        ui_layout:
          description: 'ui_layout'
          type: array
          items:
            additionalProperties: { }
        properties:
          type: array
          items:
            $ref: '#/components/schemas/PropertyDefinition'
        created:
          $ref: '#/components/schemas/Timestamp'
        updated:
          $ref: '#/components/schemas/Timestamp'
      required:
        - id
        - name
        - created
        - updated
    ThingTypeDetail:
      description: 'User with fields.'
      type: object
      properties:
        id:
          $ref: '#/components/schemas/Id'
        name:
          $ref: '#/components/schemas/Name'
        properties:
          type: array
          items:
            $ref: '#/components/schemas/PropertyDefinition'
        channels:
          type: array
          items:
            $ref: '#/components/schemas/ThingTypeChannel'
        created:
          $ref: '#/components/schemas/Timestamp'
        updated:
          $ref: '#/components/schemas/Timestamp'
      required:
        - id
        - name
        - created
        - updated
    ThingTypePage:
      description: 'Page for thing types.'
      type: object
      properties:
        content:
          description: 'List of entries on page.'
          type: array
          items:
            $ref: '#/components/schemas/ThingType'
        total_size:
          description: 'Total available entries.'
          type: integer
          format: int32
          example: 0
      required:
        - content
        - total_size
    ThingTypeCreate:
      description: 'User to create.'
      type: object
      properties:
        name:
          $ref: '#/components/schemas/Name'
        thing_type_reference:
          description: 'thing_type_reference'
          type: string
        json_schema:
          description: 'json_schema'
          type: object
          additionalProperties: { }
        ui_layout:
          description: 'ui_layout'
          type: array
          items:
            additionalProperties: { }
      required:
        - name
    ThingTypeUpdate:
      description: 'User to update.'
      type: object
      properties:
        name:
          $ref: '#/components/schemas/Name'
        thing_type_reference:
          description: 'thing_type_reference'
          type: string
        json_schema:
          description: 'json_schema'
          type: object
          additionalProperties: { }
        ui_layout:
          description: 'ui_layout'
          type: array
          items:
            additionalProperties: { }
      required:
        - name
    ThingTypeChannel:
      description: 'Thing Type Channel with fields.'
      type: object
      properties:
        id:
          $ref: '#/components/schemas/Id'
        key:
          type: string
          example: power
          description: key for the channel type
        name:
          $ref: '#/components/schemas/Name'
        description:
          type: string
          example: 1-0:1.8.0*255
          description: description for the channel type
        properties:
          type: array
          items:
            $ref: '#/components/schemas/PropertyDefinition'
        created:
          $ref: '#/components/schemas/Timestamp'
        updated:
          $ref: '#/components/schemas/Timestamp'
      required:
        - id
        - name
        - created
        - updated
...
    PropertyDefinition:
      type: object
      properties:
        name:
          $ref: '#/components/schemas/Name'
        key:
          type: string
          example: serialId
          description: key for the property
        input_type:
          type: string
          example: text
          enum: [ text, number ]
          description: input type for the frontend
...
  parameters:
    PageablePage:
      description: 'Page number for pagination.'
      name: page
      in: query
      schema:
        type: integer
        format: int32
        minimum: 0
      example: 0
    PageableSize:
      description: 'Page size for pagination.'
      name: size
      in: query
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
      example: 50
    Sort:
      description: 'Sorting.'
      name: sort
      in: query
      schema:
        type: array
        items:
          type: string
          pattern: '[a-zA-Z_]{2,10}(,(asc|desc|ASC|DESC))?'
      example:
        - name,asc
        - name,desc
    ThingTypeId:
      name: thing_type_id
      in: path
      required: true
      schema:
        type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    Keycloak:
      type: openIdConnect
      openIdConnectUrl: http://keycloak:8080/realms/inoa/.well-known/openid-configuration